2013-06-06 01:56 UTC-0800 Pritpal Bedi (bedipritpal@hotmail.com)
* contrib/hbtip/client.prg
! Fixed: to test if a connection is available actually or not
before sending other info. This fixes a RTE when a SSL over
HTTP request was made and internet access was not available.

* contrib/hbtip/httpcli.prg
+ Added: METHOD Head() for HEAD verb of HTTP interface.

METHOD SetConnectionPersistent()
This does not send "Connection: close" header entry.
Useful in cases where many requests are required to
be submitted under one session. :close() may be
called to close the connection explicitly.

METHOD IsConnectionAlive()
It allows the application to test connection's state
and facilitates to take alternate action.